+++ include/penduli.hpp
@@ -6,7 +6,7 @@
 #ifndef PENDULI_H
 #define PENDULI_H
 
-std::vector<double> classicalPendulum(std::vector<double> coord, double k);
-std::vector<double> quantumPendulum(std::vector<double> coord, double gamma);
+std::vector<long double> classicalPendulum(std::vector<long double> coord, double k);
+std::vector<long double> quantumPendulum(std::vector<long double> coord, double gamma);
 double flutuation(double x, double y, double gamma);
 #endif
