+++ include/rungekutta4thSquare.hpp
@@ -7,13 +7,13 @@
 #include<vector>
 #include<cmath>
 
-std::vector<double> updateCoord(std::vector<double>coord, std::vector<double> increment, double step, double scale,int dimension);
-std::vector<double> rungeKutta4thSquare(std::vector<double> (*function)(std::vector<double>, double), 
-                                  std::vector<double>& coord, double param, double step, 
+std::vector<long double> updateCoord(std::vector<long double>coord, std::vector<long double> increment, double step, double scale,int dimension);
+std::vector<long double> rungeKutta4thSquare(std::vector<long double> (*function)(std::vector<long double>, double), 
+                                  std::vector<long double>& coord, double param, double step, 
                                   int dimension );
-void completeRungeKuttaToFile(std::vector<double> (*function)(std::vector<double>, double), std::vector<double> initialCond,
+void completeRungeKuttaToFile(std::vector<long double> (*function)(std::vector<long double>, double), std::vector<long double> initialCond,
                                                    double param, double step, int dimension, double time_span[2]);
-std::vector<double> rungeKutta4thSquarePertubation(std::vector<double> (*function)(std::vector<double>,std::vector<double>, double), 
-                                  std::vector<double> coord,std::vector<double> pertubation, double param, double step, 
-                                  int dimension, std::vector<double>& functionAval );
+std::vector<long double> rungeKutta4thSquarePertubation(std::vector<long double> (*function)(std::vector<long double>,std::vector<long double>, double), 
+                                  std::vector<long double> coord,std::vector<long double> pertubation, double param, double step, 
+                                  int dimension, std::vector<long double>& functionAval );
 #endif
